Refactor SpeechSession into main-actor modules
